admin side page login user management done

+++ clouzie/accounts/views.py
@@ -1,6 +1,6 @@
-from django.shortcuts import render,redirect
+from django.shortcuts import render,redirect,get_object_or_404
 from django.http import HttpResponse
-from .models import CustomUser,Otp
+from .models import CustomUser,Otp,Address
 import re
 import random
 from django.utils import timezone
@@ -14,26 +14,41 @@ from django.contrib.auth import authenticate
 from django.contrib.auth import login
 from django.contrib import messages
 from django.contrib.auth import logout
+from django.contrib.auth.decorators import login_required
+from django.views.decorators.cache import never_cache
+from django.contrib.auth import update_session_auth_hash
 # Create your views here.
-
 def home(request):
+    if request.user.is_authenticated:
+        return redirect('home_main')
+    
     return render(request, 'accounts/home.html')
-
+@never_cache
 def signin(request):
+    if request.user.is_authenticated:
+        return redirect('home_main')
+    
     if request.method == 'POST':
         lemail = request.POST.get('email')
         lpassword = request.POST.get('password')
+        try:
+            user_obj = CustomUser.objects.get(email=lemail)
         
-        user = authenticate(request,email=lemail,password=lpassword)
-        if user is not None:
-            login(request,user)
-            return redirect('home_main')
+            user = authenticate(request,email=user_obj.email,password=lpassword)
+            
+            if user is not None:
+                login(request,user)
+                request.session['user_id'] = user.id
+                request.session.set_expiry(1209600)
+                return redirect('home_main')
+            return render(request,"accounts/login_page.html",{"error":"invalid email or password"})
         
-        return render(request,"accounts/login_page.html",{"error":"Incorrect login details.","email":lemail,"password":lpassword})
-        
-        
+        except CustomUser.DoesNotExist:
+            return render(request,"accounts/login_page.html",{"error":"invalid email or password"})
+          
     return render(request,"accounts/login_page.html")
 
+@never_cache
 def signup(request):
     
     if request.method == 'POST':
@@ -87,7 +102,7 @@ def signup(request):
 
         email.attach_alternative(html_content, "text/html")
         email.send()
-        
+        request.session['verify_user_id'] = user.id
         return redirect('verify')
         
     return render(request,"accounts/signup.html")
@@ -101,8 +116,12 @@ def valid_email(email):
 def valid_password(password):
     return re.fullmatch(r'^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{6,}$', password)
 
-
+@never_cache
 def verify(request):
+    user_id = request.session.get('verify_user_id')
+    if not user_id:
+        return redirect('sigin')
+    
     if request.method == "POST":
         Otp_input = ''.join([request.POST.get(f'v{i}','') for i in range(1,7)]) 
         
@@ -129,8 +148,13 @@ def verify(request):
         request.session.pop('user_id',None)
         return redirect('sigin')
     return render(request,"accounts/verify.html")
-        
+
+@never_cache  
 def resend_otp(request):
+    verify_user = request.session.get('verify_user_id')
+    if not verify_user:
+        return redirect('sigin')
+    
     user_id = request.session.get('user_id')
     Otp.objects.filter(user_id=user_id).delete()
     if not user_id:
@@ -160,7 +184,7 @@ def resend_otp(request):
     email.attach_alternative(html_content, "text/html")
     email.send()
     return JsonResponse({'success':True})
-         
+@never_cache 
 def forgot_password(request):
     if request.method == 'POST':
         femail = request.POST.get('email')
@@ -172,7 +196,7 @@ def forgot_password(request):
             return render(request, "accounts/forgot_password.html",{"error":"Email is required"})
         try:
             user = CustomUser.objects.get(email=femail)
-            request.session['user_id'] = user.id
+            request.session['forgot_user_id'] = user.id
             Otp.objects.filter(user_id=user.id).delete()
             Otp.objects.create(
                 code=otp_code,
@@ -199,10 +223,16 @@ def forgot_password(request):
         
     return render(request,"accounts/forgot_password.html")
 
+@never_cache
 def forgot_verify(request):
+    forgot_user = request.session.get('forgot_user_id')
+    if not forgot_user:
+        return redirect('sigin')
+        
     if request.method == 'POST':
+        
         Otp_input = ''.join([request.POST.get(f'v{i}','') for i in range(1,7)])
-        user_id = request.session.get('user_id')
+        user_id = request.session.get('forgot_user_id')
         otp_obj = Otp.objects.get(user_id=user_id)
         if otp_obj.code != Otp_input:
             return render(request,'accounts/forgot_verify.html',{"error":"The verification code you entered is incorrect. Please try again."})
@@ -218,7 +248,13 @@ def forgot_verify(request):
         
     return render(request,"accounts/forgot_verify.html")
 
+@never_cache
+@never_cache
 def rest_password(request):
+    forgot_user = request.session.get('forgot_user_id')
+    if not forgot_user:
+        return redirect('sigin')
+    
     if request.method == 'POST':
         rpassword = request.POST.get('password')
         cpassword = request.POST.get('cnfmpassword')
@@ -233,26 +269,27 @@ def rest_password(request):
             return render(request,"accounts/reset_password.html",{"error":"Password must be 6+ chars with letters & numbers"})
         user.set_password(rpassword)
         user.save()
+        request.session.pop('forgot_user_id',None)
         return redirect('sigin')
     return render(request,"accounts/reset_password.html")
-
+@login_required
 def main_home(request):
     return render(request,'accounts/main_page.html')
-        
-def profile(request):
-    user = request.user
-    user_details = CustomUser.objects.get(id=user.id)
-    return render(request,"accounts/profile.html",{"user":user_details})    
+@login_required()
+@never_cache
+def profile(request):  
+    return render(request,"accounts/profile.html",{"user":request.user}) 
 
+@login_required
+@never_cache
 def change_password(request):
+    
     if request.method == "POST":
         user = request.user
         old_password = request.POST.get('old_password')
         new_password = request.POST.get('new_password')
         cnfrm_password = request.POST.get('confirm_password')
-        print(old_password)
-        print(new_password)
-        print(cnfrm_password)
+
         if not user.check_password(old_password):
             return render(request,"accounts/change_password.html",{'error':"Old passowrd incorrect Try again.","old_password":old_password,"new_password":new_password,"cnfrm_password":cnfrm_password})
         
@@ -265,11 +302,14 @@ def change_password(request):
         
         user.set_password(new_password)
         user.save()
-        return HttpResponse("password changed now ")
+        update_session_auth_hash(request,user)
+        return redirect('profile')
     
     return render(request,'accounts/change_password.html') 
 
 
+@login_required
+@never_cache
 def edit_profile(request):
     user = request.user
     user_details = CustomUser.objects.get(id=user.id)
@@ -277,8 +317,9 @@ def edit_profile(request):
         username = request.POST.get('name')
         phone = request.POST.get('phone')
         image = request.FILES.get('profile_image')
-        if user.profile_photo:
-         user.profile_photo.delete(save=False)
+        if image is not None:
+            if user.profile_photo:
+                user.profile_photo.delete(save=False)
         
         user.username = username
         user.phone = phone
@@ -289,6 +330,9 @@ def edit_profile(request):
         return redirect('profile')
     return render(request,"accounts/edit_profile.html",{"user":user_details})
 
+
+@login_required
+@never_cache
 def remove_profile(request):
     user = request.user
     if user.profile_photo:
@@ -301,4 +345,83 @@ def remove_profile(request):
 def logout_page(request):
     logout(request)
     return redirect('home')
+
+def dummy(request):
+    return HttpResponse("hello world")   
+@login_required
+@never_cache
+def adress(request):
+    address = Address.objects.filter(user=request.user).order_by('-is_default')
+    return render(request,"accounts/address.html",{"address":address})
+@login_required
+@never_cache
+def add_address(request):
+    if request.method == "POST":
+        full_name = request.POST.get('full_name')
+        phone_number = request.POST.get('phone_number')
+        pincode = request.POST.get('pincode')
+        city = request.POST.get('city') 
+        state = request.POST.get('state') 
+        address_line1 = request.POST.get('address_line1') 
+        address_line2 = request.POST.get('address_line2') 
+        is_default = bool(request.POST.get('is_default'))
+        address_type = request.POST.get('type')
         
+        if is_default:
+            Address.objects.filter(user=request.user).update(is_default=False)
+            
+        fields = [
+        "full_name", "phone_number", "address_line1",
+        "city", "state", "pincode"
+        ]
+
+        for field in fields:
+            if not request.POST.get(field):
+                return redirect("/add-address/?error=1")
+        Address.objects.create(
+            user=request.user,
+            full_name=full_name,
+            phone_number=phone_number,
+            pincode=pincode,
+            city=city,
+            state=state,
+            address_line1=address_line1,
+            address_line2=address_line2,
+            is_default = is_default,
+            type = address_type   
+        )
+        return redirect('/address/?success=1')
+    return render(request,"accounts/add_address.html")
+@login_required
+@never_cache
+def edit_address(request,id):
+    address = get_object_or_404(Address,id=id,user=request.user)
+    if request.method == "POST":
+        is_default = bool(request.POST.get('is_default'))
+        address.full_name = request.POST.get('full_name')
+        address.phone_number = request.POST.get('phone_number')
+        address.pincode = request.POST.get('pincode')
+        address.city = request.POST.get('city') 
+        address.state = request.POST.get('state') 
+        address.address_line1 = request.POST.get('address_line1') 
+        address.address_line2 = request.POST.get('address_line2') 
+        address.is_default = bool(request.POST.get('is_default'))
+        address.address_type = request.POST.get('type')
+        
+        if is_default:
+            Address.objects.filter(user=request.user).update(is_default=False)
+        address.save()
+        return redirect('address')
+    return render(request,"accounts/edit_address.html",{"address":address})
+@login_required
+@never_cache
+def delete_address(request,id):
+    address = get_object_or_404(Address,id=id)
+    if address.is_default:
+        new = Address.objects.filter(user=request.user).exclude(id=id).first()
+        if new:
+            new.is_default = True
+            new.save()
+        
+    address.delete()
+    return redirect('address')
